Fix virtual host selection by wrapping `ServerTlsChannel`
The previous implementation didn't correctly dispatch on different
virtual hosts because `SniKeyManager.chooseEngineServerAlias` isn't
ever called when using an `SSLServerSocket`. It is only used when
directly working with `SSLEngine`.

`tls-channel` does the hard part of setting up a TLS channel, and the
remaining `TLSUtils` functions just need to load the keys and create
`SSLContext` per host.